Hooked up the main view, I still need to delete some labels and fixed some functionalities
